pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yuki-conductor"
version = "0.1.0"
description = "Bridge Slack messages to Claude Code CLI"
requires-python = ">=3.11"
dependencies = [
    "slack-bolt>=1.18.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "croniter>=1.4.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.9",
]

[project.scripts]
yuki-conductor = "yuki_conductor.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/yuki_conductor"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]
